Final update

+++ assignment2023cpp-throwmeister-main/Part4/SetSolver.h
@@ -96,7 +96,7 @@ public:
         }
     }
 
-    int ReturnValue(size_t row, size_t col){
+    int ReturnValue(int row, int col){
         return board[row][col].readValue;
     }
